Update delta_iter.py
It was a known issue that some valuations were not skipped in the previous version of deltaiter (although this doe snot affect results, it affects the running time as we are checking valuations that could be skipped). This was due to the fact that lists of deltas whose value at the largest index is 0 were not tested against the valuations created by a plus_one with larger index. E.g. when moving from [0,1,2,2] (all max values are 3) to [0,2,0,0], the list of deltas [(2,1),(0,2)] was not tested. The current commit fixes this.